feat: Don't return passwords when fetching user data
